requirements.txt
about-time==4.2.1
alive-progress==3.0.1
ase==3.22.1
asttokens==2.2.1
backcall==0.2.0
bayesian-optimization==1.4.2
certifi==2022.9.24
charset-normalizer==2.1.1
colorama==0.4.6
comm==0.1.2
contourpy==1.0.6
cycler==0.11.0
debugpy==1.6.6
decorator==5.1.1
emmet-core==0.38.9
executing==1.2.0
extxyz==0.1.2
fonttools==4.38.0
future==0.18.2
grapheme==0.6.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.20.2
ipython==8.9.0
jarvis-tools==2023.1.8
jedi==0.18.2
joblib==1.2.0
jupyter-client==8.0.2
jupyter-core==5.2.0
kiwisolver==1.4.4
lammps==2022.6.23
latexcodec==2.0.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
monty==2022.9.9
mp-api==0.29.7
mpi4py==3.1.4
mpmath==1.2.1
msgpack==1.0.4
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.5
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
packaging==21.3
palettable==3.3.0
pandas==1.5.1
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.3.0
platformdirs==2.6.2
plotly==5.11.0
prompt-toolkit==3.0.36
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pybtex==0.24.0
pydantic==1.10.2
Pygments==2.14.0
pyleri==1.4.1
pymatgen==2022.11.7
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.6
PyYAML==6.0
pyzmq==25.0.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.1
scipy==1.9.3
six==1.16.0
spglib==2.0.2
stack-data==0.6.2
sympy==1.11.1
tabulate==0.9.0
tenacity==8.1.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==1.13.0
tornado==6.2
tqdm==4.64.1
traitlets==5.9.0
typing-extensions==4.4.0
uncertainties==3.1.7
urllib3==1.26.12
wcwidth==0.2.6
xmltodict==0.13.0
zipp==3.12.0
